requirements.txt
requests
gevent
m3u8
selenium
seleniumwire
python-dotenv